refactor embeddings writer
