+++ src/eva/core/callbacks/writers/embeddings/__init__.py
@@ -1,0 +1,5 @@
+"""Embedding callback writers."""
+
+from eva.core.callbacks.writers.embeddings.classification import ClassificationEmbeddingsWriter
+
+__all__ = ["ClassificationEmbeddingsWriter"]
