+++ src/eva/core/callbacks/writers/embeddings/typings.py
