+++ tests/eva/core/callbacks/writers/embeddings/__init__.py
@@ -1,0 +1,1 @@
+"""Embeddings writer callback unit tests."""
